add common LineMap
